Ajout de a structure du lucky block
