+++ src/main/java/fr/communaywen/core/luckyblocks/events/SkeletonHorseEvent.java
@@ -1,0 +1,38 @@
+package fr.communaywen.core.luckyblocks.events;
+
+import org.bukkit.Location;
+import org.bukkit.Material;
+import org.bukkit.block.Block;
+import org.bukkit.block.BlockState;
+import org.bukkit.block.Skull;
+import org.bukkit.entity.Player;
+import org.bukkit.entity.SkeletonHorse;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockBreakEvent;
+
+public class SkeletonHorseEvent implements Listener {
+
+    @EventHandler
+    public void onBreak(BlockBreakEvent e) {
+
+        Player player = e.getPlayer();
+        Location playerLocation = player.getLocation();
+        Block block = e.getBlock();
+        BlockState bs = block.getState();
+        Location location = e.getBlock().getLocation().add(0.5, 0, 0.5);
+
+
+        if (bs instanceof Skull) {
+
+            Skull skull = (Skull) bs;
+            if (skull.getOwner().equalsIgnoreCase("luck")) {
+                e.setCancelled(true);
+                block.setType(Material.AIR);
+
+                location.getWorld().spawn(location, SkeletonHorse.class);
+
+                }
+            }
+        }
+    }
